Deprecate Old Pathology Components (#4415)
* Deprecate pathology datasets


* Update min test exclusion


* Update test file names for deprecated component and update prob map deprecation
